3 Find
Here we will learn how to search for documents.

For all of the
exercises, the database is learnyoumongo.
So, the url would be something
like: mongodb://localhost:27017/learnyoumongo

Use the parrots
collection to find all documents where age is greater than the first
argument passed to your script.

Using console.log, print the documents
to stdout.