4 Find project
Here we will learn how to search for documents but only fetch the fields
we need. Also known as projection in MongoDB

Use the parrots collection
to find all documents where age is greater than the first argument
passed to your script.

The difference from the last lesson will be that
we only want the name and age properties

Using console.log, print the
documents to stdout.